feat: add InternalMoveLineNotes component
